finished result dao and answer dao
Former-commit-id: ed882546da318249462773d61a39fc8873288876
Former-commit-id: 50eac26a259b1f0ab905a122147404a1a3a1467c [formerly 75bb841d6f25e6b61e0ed58b88495a11b493fb96]
Former-commit-id: 081a53bdd29c93c0deb94ec3b6603027486a28f6

+++ src/mitll/langtest/server/ScoreServlet.java
@@ -394,7 +394,7 @@ public class ScoreServlet extends DatabaseServlet {
 
       //logger.debug("chapterHistory " + user + " selection " + selection);
       try {
-        long l = Long.parseLong(userAndSelection.getUser());
+        int l = Integer.parseInt(userAndSelection.getUser());
         toReturn = db.getJsonScoreHistory(l, selection, getExerciseSorter());
       } catch (NumberFormatException e) {
         toReturn.put(ERROR, "User id should be a number");
@@ -737,7 +737,7 @@ public class ScoreServlet extends DatabaseServlet {
         if (doFlashcard) {
           jsonForScore.put(IS_CORRECT, answer.isCorrect());
           jsonForScore.put(SAID_WORD, answer.isSaidAnswer());
-          long decodeResultID = answer.getResultID();
+          int decodeResultID = answer.getResultID();
           jsonForScore.put("resultID", decodeResultID);
 
           // attempt to get more feedback when we're too sensitive and match the unknown model
